Separated functionality into logical files
I was originally trying to keep this one single file because I didn't
want to have the program be a directory, but once I introduced the
dependency on ply, I decided that if it was no longer self-contained and
totally core functionality, I might as well break it out into logical
moudules. Also, the neigh one-thousand-line logrok.py was getting
unweildy.